+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 ### Unreleased
+
+### 2.1.4
+- Upgrade twreporter-react-index-page-components to 1.1.0
+- Update tool styles, remove raf, and check screen type in Article
+- Remove lazyload on photograph articles
+- Update ref to callback function
+- Replace pure-render-mixin with PureComponent
+
 ### 2.1.3
 - Bug fix. Fix og:image missing
 
@@ -63,6 +71,12 @@ author and author-articles reducers.
 - remove /api folders
 - update unit test
 - remove bin/api.js
+
+### 1.1.19
+- [Update] Update the navigation bar
+
+### 1.1.20
+- [Update] Change the homepage banner
 
 ### 1.1.19
 - [Update] Update the navigation bar